requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
uvicorn[standard]
openai==1.12.0
python-dotenv==1.0.1
websockets==12.0